requirements.txt
fpdf==1.7.2
fpdf2==2.6.1
pandas==1.4.4
pdf2image==1.16.0
Pillow==9.4.0
streamlit==1.19.0
streamlit_authenticator